requirements.txt
# requirements.txt

# Backend Dependencies
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
sqlparse==0.5.3
typing_extensions==4.13.2
certifi==2025.7.14
python-decouple==3.8